fix(release.config.js): trying a fix for Semantic-Release Config
This might fix the issue of Semantic-Release failing to push Git Tag with GITHUB_TOKEN. We removed
the 'repositoryUrl' with value as ssh git repo link from release.config.js - as it was needed to
test locally but it is creating a problem maybe on CI/CD. Let's see if it does fix it. If it does,
then good.

fix #1